+++ internal/server/dispatcher/response.go
@@ -127,8 +127,10 @@ func (d *Dispatcher) writeResponse(w http.ResponseWriter, r *http.Request, opera
 				extractCloudFrontETag(w, resp, payloadRoot)
 				err = protocol.EncodeRestXMLPayloadResponse(w, payloadRoot, resp)
 			} else {
-				err = protocol.EncodeRestXMLResponse(w, opName, resp)
+				err = protocol.EncodeRestXMLResponseWithNamespace(w, opName, "https://cloudfront.amazonaws.com/doc/2020-05-31/", resp)
 			}
+		} else if strings.HasPrefix(r.URL.Path, "/2013-04-01/") {
+			err = protocol.EncodeRestXMLResponse(w, opName, resp)
 		} else {
 			err = protocol.EncodeRestXMLResponse(w, opName, resp)
 		}